fix(site): add missing OG image for /compute page
The compute page metadata references /og/og-compute.png, but the file
was never committed, so the OG image 404s in production. Serve it from
a route handler that renders the image at build time, and extract the
OG template shared with the blog OG route into src/lib/og-image.tsx.
